Make movie lists general and return movielist
this change is just in serverless part

+++ Api/Functions/MovieLists.cs
@@ -14,7 +14,7 @@ using MovieFiles.Core.Interfaces;
 using System;
 using System.Collections.Generic;
 using MovieFiles.Core.Models;
-
+using System.Linq;
 
 namespace MovieFiles.Api.Functions
 {
@@ -22,17 +22,20 @@ namespace MovieFiles.Api.Functions
     {
         private readonly ILogger<MovieLists> _logger;
         private readonly IMovieListRepository _movieListService;
+        private readonly IMovieDetailsService _movieDetailService;
 
-        public MovieLists(ILogger<MovieLists> log, IMovieListRepository movieListService)
+        public MovieLists(ILogger<MovieLists> log, IMovieListRepository movieListService, IMovieDetailsService movieDetailsService)
         {
             _logger = log;
             _movieListService = movieListService;
+            _movieDetailService = movieDetailsService;
         }
 
-        [FunctionName("AddMovieToWatchLater")]
-        [OpenApiOperation(operationId: "AddMovieToWatchLater", tags: new[] { "Movie lists" })]
+        [FunctionName("AddMovieToMovieList")]
+        [OpenApiOperation(operationId: "AddMovieToMovieList", tags: new[] { "Movie lists" })]
         [OpenApiParameter(name: "userId", In = ParameterLocation.Query, Required = true, Type = typeof(Guid), Description = "Id of user that is adding movie to his list")]
         [OpenApiParameter(name: "movieId", In = ParameterLocation.Query, Required = true, Type = typeof(int), Description = "Id of movie to add to the list")]
+        [OpenApiParameter(name: "movieListType", In = ParameterLocation.Query, Required = true, Type = typeof(MyMovieListItem.ListType), Description = "Type of movie list. 0: watch later, 1: toplist, 2: already watched")]
         [OpenApiParameter(name: "x-functions-key", In = ParameterLocation.Header, Required = true, Type = typeof(string), Description = "The function key")]
         [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "text/plain", bodyType: typeof(string), Description = "The OK response")]
         [OpenApiResponseWithBody(statusCode: HttpStatusCode.BadRequest, contentType: "text/plain", bodyType: typeof(string), Description = "Incorrect parameters were provided.")]
@@ -47,9 +50,16 @@ namespace MovieFiles.Api.Functions
             if (!Guid.TryParse(req.Query["userId"], out var userId)){
                 return new BadRequestObjectResult("Invalid user ID.");
             }
+            if (!Enum.TryParse<MyMovieListItem.ListType>(req.Query["movieListType"],true,out var movieListType)){
+                return new BadRequestObjectResult("Invalid type of list.");
+            }
 
             var additionSuccesful = await _movieListService.AddMovieToMyList(
-                MyMovieListItem.CreateWatchLaterListItem(movieId,userId)
+                new(){
+                    UserId = userId,
+                    MovieId = movieId,
+                    ListName = MyMovieListItem.GetListTypeName(movieListType)
+                }
             );
 
             if (additionSuccesful){
@@ -61,12 +71,13 @@ namespace MovieFiles.Api.Functions
             
         }
 
-        [FunctionName("GetMoviesToWatchLater")]
-        [OpenApiOperation(operationId: "GetMoviesToWatchLater", tags: new[] { "Movie lists" })]
+        [FunctionName("GetMoviesFromMovieList")]
+        [OpenApiOperation(operationId: "GetMoviesFromMovieList", tags: new[] { "Movie lists" })]
         [OpenApiParameter(name: "userId", In = ParameterLocation.Query, Required = true, Type = typeof(Guid), Description = "Id of a user to get the list of.")]
+        [OpenApiParameter(name: "movieListType", In = ParameterLocation.Query, Required = true, Type = typeof(MyMovieListItem.ListType), Description = "Type of movie list. 0: watch later, 1: toplist, 2: already watched")]
         [OpenApiParameter(name: "page", In = ParameterLocation.Query, Required = true, Type = typeof(int), Description = "Page number to return.")]
         [OpenApiParameter(name: "x-functions-key", In = ParameterLocation.Header, Required = true, Type = typeof(string), Description = "The function key")]
-        [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "application/json", bodyType: typeof(CustomMovieList<MyMovieListItem>), Description = "The OK response")]
+        [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "application/json", bodyType: typeof(Core.Models.MovieList), Description = "The OK response")]
         [OpenApiResponseWithBody(statusCode: HttpStatusCode.BadRequest, contentType: "text/plain", bodyType: typeof(string), Description = "Incorrect parameters were provided.")]
         public async Task<IActionResult> GetMoviesToWatchLater(
             [HttpTrigger(AuthorizationLevel.Function, "get", Route = "my-movies/watch-later")] HttpRequest req)
@@ -80,16 +91,20 @@ namespace MovieFiles.Api.Functions
             if (!int.TryParse(req.Query["page"], out var page) && page <1){
                 return new BadRequestObjectResult("Invalid page number.");
             }
+            if (!Enum.TryParse<MyMovieListItem.ListType>(req.Query["movieListType"],true,out var movieListType)){
+                return new BadRequestObjectResult("Invalid type of list.");
+            }
 
-            CustomMovieList<MyMovieListItem> list = await _movieListService.GetMyMovieList(userId, MyMovieListItem.ListType.WATCH_LATER, page);
+            CustomMovieList<MyMovieListItem> list = await _movieListService.GetMyMovieList(userId, movieListType, page);
 
-            return new OkObjectResult(list);
+            return new OkObjectResult(await ConvertCustomMovieListToMovieList(list));
         }
 
-        [FunctionName("RemoveMovieToWatchLater")]
-        [OpenApiOperation(operationId: "RemoveMovieToWatchLater", tags: new[] { "Movie lists" })]
+        [FunctionName("RemoveMovieFromMovieList")]
+        [OpenApiOperation(operationId: "RemoveMovieFromMovieList", tags: new[] { "Movie lists" })]
         [OpenApiParameter(name: "userId", In = ParameterLocation.Query, Required = true, Type = typeof(Guid), Description = "Id of user that is adding movie to his list")]
         [OpenApiParameter(name: "movieId", In = ParameterLocation.Query, Required = true, Type = typeof(int), Description = "Id of movie to add to the list")]
+        [OpenApiParameter(name: "movieListType", In = ParameterLocation.Query, Required = true, Type = typeof(MyMovieListItem.ListType), Description = "Type of movie list. 0: watch later, 1: toplist, 2: already watched")]
         [OpenApiParameter(name: "x-functions-key", In = ParameterLocation.Header, Required = true, Type = typeof(string), Description = "The function key")]
         [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "text/plain", bodyType: typeof(string), Description = "The OK response")]
         [OpenApiResponseWithBody(statusCode: HttpStatusCode.BadRequest, contentType: "text/plain", bodyType: typeof(string), Description = "Incorrect parameters were provided.")]
@@ -104,9 +119,16 @@ namespace MovieFiles.Api.Functions
             if (!Guid.TryParse(req.Query["userId"], out var userId)){
                 return new BadRequestObjectResult("Invalid user ID.");
             }
+            if (!Enum.TryParse<MyMovieListItem.ListType>(req.Query["movieListType"],true,out var movieListType)){
+                return new BadRequestObjectResult("Invalid type of list.");
+            }
 
             var deletionSuccesfull = await _movieListService.RemoveMovieFromMyList(
-                MyMovieListItem.CreateWatchLaterListItem(movieId,userId)
+                new(){
+                    UserId = userId,
+                    MovieId = movieId,
+                    ListName = MyMovieListItem.GetListTypeName(movieListType)
+                }
             );
 
             if (deletionSuccesfull){
@@ -114,7 +136,23 @@ namespace MovieFiles.Api.Functions
             } else {
                 return new NotFoundObjectResult("Unable to find resource to delete");
             }
-            
+        }
+
+        private async Task<Core.Models.MovieList> ConvertCustomMovieListToMovieList(CustomMovieList<Core.Models.MyMovieListItem> externList)
+        {
+            Core.Models.MovieList myList = new()
+            {
+                Page = externList.Page,
+                TotalResults = externList.TotalResults,
+                TotalPages = externList.TotalPages
+            };
+            var getMoviesTasks = externList.List.Select(getMovieFromListItem).ToArray();
+            myList.Results = await Task.WhenAll(getMoviesTasks);
+            return myList;
+        }
+
+        private async Task<Core.Models.Movie> getMovieFromListItem(Core.Models.MyMovieListItem item){
+            return await _movieDetailService.GetMovieDetailsAsync(item.MovieId);
         }
     }
 }
